tools/istio-iptables: make the validator robust with platform-specific builds
